# specs/001-todo-fullstack-webapp/contracts/openapi.yaml
openapi: 3.0.0
info:
  title: Todo API
  description: REST API for Todo Web Application
  version: 1.0.0
  servers:
    - url: http://localhost:8000/api

paths:
  /auth/register:
    post:
      summary: Register a new user
      description: Creates a new user account and returns authentication token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  maxLength: 255
                  description: User email address
                password:
                  type: string
                  minLength: 8
                  maxLength: 255
                  description: User password (minimum 8 characters)
              example:
                email: user@example.com
                password: securepassword123
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    description: JWT authentication token
                  token_type:
                    type: string
                    example: bearer
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      email:
                        type: string
                        format: email
                      created_at:
                        type: string
                        format: date-time
        '400':
          description: Validation error or email already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: Email already registered

  /auth/login:
    post:
      summary: Login user
      description: Authenticates user and returns JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  token_type:
                    type: string
                    example: bearer
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                      email:
                        type: string
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: Invalid email or password

  /{user_id}/tasks:
    get:
      summary: List user tasks
      description: Returns all tasks belonging to the authenticated user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: User ID (from JWT token)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      format: uuid
                    user_id:
                      type: string
                      format: uuid
                    title:
                      type: string
                      maxLength: 255
                    completed:
                      type: boolean
                    created_at:
                      type: string
                      format: date-time
                    updated_at:
                      type: string
                      format: date-time
        '401':
          description: Missing or invalid token
        '403':
          description: User ID mismatch (trying to access another user's data)

    post:
      summary: Create task
      description: Creates a new task for the authenticated user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 255
              example:
                title: Buy groceries
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  user_id:
                    type: string
                    format: uuid
                  title:
                    type: string
                  completed:
                    type: boolean
                    default: false
                  created_at:
                    type: string
                    format: date-time
                  updated_at:
                    type: string
                    format: date-time
        '401':
          description: Missing or invalid token
        '422':
          description: Validation error (empty title)

  /{user_id}/tasks/{task_id}:
    get:
      summary: Get task by ID
      description: Returns a specific task if owned by authenticated user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: task_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Task found
        '404':
          description: Task not found
        '403':
          description: Task belongs to another user

    put:
      summary: Update task
      description: Updates the title of a task
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: task_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 255
      responses:
        '200':
          description: Task updated successfully
        '404':
          description: Task not found
        '422':
          description: Validation error

    delete:
      summary: Delete task
      description: Deletes a task
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: task_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Task deleted successfully
        '404':
          description: Task not found

  /{user_id}/tasks/{task_id}/complete:
    patch:
      summary: Toggle task completion
      description: Toggles the completion status of a task
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: task_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - completed
              properties:
                completed:
                  type: boolean
              example:
                completed: true
      responses:
        '200':
          description: Task completion status updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  completed:
                    type: boolean
        '404':
          description: Task not found

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT